+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageGrpcWriteChannelTest.java
@@ -2,12 +2,13 @@ package com.google.cloud.hadoop.gcsio;
 
 import static com.google.cloud.hadoop.util.AsyncWriteChannelOptions.PIPE_BUFFER_SIZE_DEFAULT;
 import static com.google.common.truth.Truth.assertThat;
+import static com.google.storage.v2.ServiceConstants.Values.MAX_WRITE_CHUNK_BYTES;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.atLeast;
 import static org.mockito.Mockito.eq;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
@@ -36,6 +37,7 @@ import com.google.storage.v2.WriteObjectRequest;
 import com.google.storage.v2.WriteObjectResponse;
 import com.google.storage.v2.WriteObjectSpec;
 import io.grpc.ManagedChannelBuilder;
+import io.grpc.Server;
 import io.grpc.Status;
 import io.grpc.StatusException;
 import io.grpc.inprocess.InProcessChannelBuilder;
@@ -46,6 +48,7 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.ClosedChannelException;
 import java.time.Duration;
+import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -79,6 +82,13 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
                           .setBucket(BUCKET_NAME)
                           .setName(OBJECT_NAME)
                           .setContentType(CONTENT_TYPE)))
+          .build();
+  private static final Object DEFAULT_OBJECT =
+      Object.newBuilder()
+          .setBucket(BUCKET_NAME)
+          .setName(OBJECT_NAME)
+          .setGeneration(1)
+          .setMetageneration(2)
           .build();
   private static final QueryWriteStatusRequest WRITE_STATUS_REQUEST =
       QueryWriteStatusRequest.newBuilder().setUploadId(UPLOAD_ID).build();
@@ -134,8 +144,7 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
             .build();
 
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
-    verify(fakeService.insertRequestObserver, times(1)).onNext(expectedInsertRequest);
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    verifyInsertRequestObservers(expectedInsertRequest);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 2);
   }
 
@@ -160,8 +169,7 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
             .build();
 
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
-    verify(fakeService.insertRequestObserver, times(1)).onNext(expectedInsertRequest);
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    verifyInsertRequestObservers(expectedInsertRequest);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 2);
   }
 
@@ -183,12 +191,8 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     writeChannel.write(data.asReadOnlyByteBuffer());
     writeChannel.close();
 
-    ArgumentCaptor<WriteObjectRequest> requestCaptor =
-        ArgumentCaptor.forClass(WriteObjectRequest.class);
-
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
-    verify(fakeService.insertRequestObserver, times(1)).onNext(requestCaptor.capture());
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    verifyInsertRequestObservers(null);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 2);
   }
 
@@ -214,12 +218,8 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     writeChannel.write(data.asReadOnlyByteBuffer());
     writeChannel.close();
 
-    ArgumentCaptor<WriteObjectRequest> requestCaptor =
-        ArgumentCaptor.forClass(WriteObjectRequest.class);
-
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
-    verify(fakeService.insertRequestObserver, times(1)).onNext(requestCaptor.capture());
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    verifyInsertRequestObservers(null);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 2);
   }
 
@@ -238,14 +238,10 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     writeChannel.write(data.asReadOnlyByteBuffer());
     writeChannel.close();
 
-    ArgumentCaptor<WriteObjectRequest> requestCaptor =
-        ArgumentCaptor.forClass(WriteObjectRequest.class);
-
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
     // TODO(b/150892988): Use this mock when implement resuming after a transient error.
     // verify(fakeService, times(1)).queryWriteStatus(eq(WRITE_STATUS_REQUEST), any());
-    verify(fakeService.insertRequestObserver, times(1)).onNext(requestCaptor.capture());
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    verifyInsertRequestObservers(null);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 2);
   }
 
@@ -389,6 +385,36 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
   }
 
   @Test
+  public void writeMoreThanRequestBufferCapacity() throws Exception {
+    AsyncWriteChannelOptions options =
+        AsyncWriteChannelOptions.builder().setNumberOfBufferedRequests(1).build();
+    ObjectWriteConditions writeConditions = ObjectWriteConditions.NONE;
+    GoogleCloudStorageGrpcWriteChannel writeChannel =
+        newWriteChannel(
+            options,
+            writeConditions,
+            /* requesterPaysProject= */ null,
+            () -> BackOff.ZERO_BACKOFF,
+            false);
+    fakeService.setQueryWriteStatusResponses(
+        ImmutableList.of(QueryWriteStatusResponse.newBuilder().setPersistedSize(1).build())
+            .iterator());
+    ByteString chunk = createTestData(MAX_WRITE_CHUNK_BYTES.getNumber() * 2 - 1);
+
+    writeChannel.initialize();
+    writeChannel.write(chunk.asReadOnlyByteBuffer());
+    writeChannel.close();
+
+    verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
+    verify(fakeService, times(0)).queryWriteStatus(eq(WRITE_STATUS_REQUEST), any());
+    assertEquals(fakeService.insertRequestObserverList.size(), 2);
+    verifyInsertRequestObservers(null);
+    // 1. One for startResumable Request
+    // 2. Two calls for creating 2 write object calls
+    headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 3);
+  }
+
+  @Test
   public void writeOneChunkWithSingleErrorAndResume() throws Exception {
     AsyncWriteChannelOptions options =
         AsyncWriteChannelOptions.builder().setUploadChunkSize(GCS_MINIMUM_CHUNK_SIZE).build();
@@ -405,9 +431,7 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     fakeService.setQueryWriteStatusResponses(
         ImmutableList.of(QueryWriteStatusResponse.newBuilder().setPersistedSize(1).build())
             .iterator());
-    ByteString chunk = createTestData(GCS_MINIMUM_CHUNK_SIZE);
-    ArgumentCaptor<WriteObjectRequest> requestCaptor =
-        ArgumentCaptor.forClass(WriteObjectRequest.class);
+    ByteString chunk = createTestData(GCS_MINIMUM_CHUNK_SIZE * 2);
 
     writeChannel.initialize();
     writeChannel.write(chunk.asReadOnlyByteBuffer());
@@ -415,10 +439,8 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
 
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
     verify(fakeService, times(1)).queryWriteStatus(eq(WRITE_STATUS_REQUEST), any());
-    verify(fakeService.insertRequestObserver, atLeast(1)).onNext(requestCaptor.capture());
-    // TODO(hgong): Figure out a way to check the expected requests and actual reqeusts builder.
-    // assertEquals(expectedRequests, requestCaptor.getAllValues());
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    assertEquals(fakeService.insertRequestObserverList.size(), 2);
+    verifyInsertRequestObservers(null);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 4);
   }
 
@@ -509,8 +531,8 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     };
     byte[] expectedCrc32C = {51, 121, -76, -54};
 
-    fakeService.setObject(
-        FakeService.DEFAULT_OBJECT
+    fakeService.setResponse(
+        DEFAULT_OBJECT
             .toBuilder()
             .setSize(9)
             .setGeneration(1)
@@ -552,6 +574,51 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
             new VerificationAttributes(expectedMd5Hash, expectedCrc32C));
 
     assertThat(itemInfo).isEqualTo(expectedItemInfo);
+  }
+
+  @Test
+  public void verifyResponseObserverThrows() throws IOException {
+    Throwable fakeException = new RuntimeException("ResponseObserver's onError is called");
+    fakeService =
+        spy(
+            new FakeService() {
+              @Override
+              public StreamObserver<WriteObjectRequest> writeObject(
+                  StreamObserver<WriteObjectResponse> responseObserver) {
+                InsertRequestObserver insertRequestObserverLatest =
+                    spy(
+                        new InsertRequestObserver(null, null) {
+                          @Override
+                          public void onCompleted() {
+                            // onCompleted on Fake RequestObserver will make sure to trigger
+                            // responseObserver's onError
+                            this.onError(fakeException);
+                          }
+                        });
+                insertRequestObserverLatest.responseObserver = responseObserver;
+                return insertRequestObserverLatest;
+              }
+            });
+
+    String serverName = InProcessServerBuilder.generateName();
+    Server server =
+        InProcessServerBuilder.forName(serverName)
+            .directExecutor()
+            .addService(fakeService)
+            .build()
+            .start();
+    stub =
+        StorageGrpc.newStub(InProcessChannelBuilder.forName(serverName).directExecutor().build());
+
+    GoogleCloudStorageGrpcWriteChannel writeChannel = newWriteChannel();
+
+    ByteString data = ByteString.copyFromUtf8("test data");
+    writeChannel.initialize();
+    writeChannel.write(data.asReadOnlyByteBuffer());
+    assertThrows(IOException.class, writeChannel::close);
+
+    verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
+    server.shutdown();
   }
 
   @Test
@@ -643,8 +710,7 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
             .build();
 
     verify(fakeService, times(1)).startResumableWrite(eq(START_REQUEST), any());
-    verify(fakeService.insertRequestObserver, times(1)).onNext(expectedInsertRequest);
-    verify(fakeService.insertRequestObserver, atLeast(1)).onCompleted();
+    verifyInsertRequestObservers(expectedInsertRequest);
     headerInterceptor.verifyAllRequestsHasGoogRequestParamsHeader(V1_BUCKET_NAME, 2);
   }
 
@@ -716,6 +782,20 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     return ByteString.copyFrom(result);
   }
 
+  private void verifyInsertRequestObservers(WriteObjectRequest expectedInsertRequest) {
+    ArgumentCaptor<WriteObjectRequest> requestCaptor =
+        ArgumentCaptor.forClass(WriteObjectRequest.class);
+    for (InsertRequestObserver requestObserver : fakeService.insertRequestObserverList) {
+      verify(requestObserver, times(1))
+          .onNext(expectedInsertRequest != null ? expectedInsertRequest : requestCaptor.capture());
+      if (requestObserver.isErrored()) {
+        verify(requestObserver, times(1)).onError(any());
+      } else {
+        verify(requestObserver, times(1)).onCompleted();
+      }
+    }
+  }
+
   private static class FakeGrpcDecorator implements StorageStubProvider.GrpcDecorator {
 
     @Override
@@ -743,15 +823,10 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
 
   private static class FakeService extends StorageImplBase {
 
-    static final Object DEFAULT_OBJECT =
-        Object.newBuilder()
-            .setBucket(BUCKET_NAME)
-            .setName(OBJECT_NAME)
-            .setGeneration(1)
-            .setMetageneration(2)
-            .build();
-
-    InsertRequestObserver insertRequestObserver = spy(new InsertRequestObserver());
+    WriteObjectResponse writeObjectResponse = null;
+    Throwable insertRequestException = null;
+    // There could be multiple requestObservers created over the course of whole file upload
+    List<InsertRequestObserver> insertRequestObserverList = new ArrayList<>();
 
     private Throwable startRequestException;
     private List<Throwable> insertObjectExceptions;
@@ -788,22 +863,23 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     @Override
     public StreamObserver<WriteObjectRequest> writeObject(
         StreamObserver<WriteObjectResponse> responseObserver) {
+      InsertRequestObserver insertRequestObserverLatest =
+          spy(new InsertRequestObserver(writeObjectResponse, insertRequestException));
+      insertRequestObserverList.add(insertRequestObserverLatest);
       if (insertObjectExceptions != null && insertObjectExceptions.size() > 0) {
         Throwable throwable = insertObjectExceptions.remove(0);
         if (!throwable.getClass().isAssignableFrom(Throwable.class)
             || throwable.getCause() != null) {
-          insertRequestObserver.insertRequestException = throwable;
-          insertRequestObserver.resumeFromInsertException = true;
+          insertRequestObserverLatest.insertRequestException = throwable;
+          insertRequestObserverLatest.resumeFromInsertException = true;
         }
       }
-      insertRequestObserver.responseObserver = responseObserver;
-      return insertRequestObserver;
+      insertRequestObserverLatest.responseObserver = responseObserver;
+      return insertRequestObserverLatest;
     }
 
-    public void setObject(Object object) {
-      this.insertRequestObserver.object = object;
-      this.insertRequestObserver.writeObjectResponse =
-          WriteObjectResponse.newBuilder().setResource(object).build();
+    public void setResponse(Object object) {
+      writeObjectResponse = WriteObjectResponse.newBuilder().setResource(object).build();
     }
 
     void setQueryWriteStatusResponses(Iterator<QueryWriteStatusResponse> responses) {
@@ -819,7 +895,7 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
     }
 
     void setInsertRequestException(Throwable t) {
-      insertRequestObserver.insertRequestException = t;
+      insertRequestException = t;
     }
 
     public void setInsertObjectExceptions(List<Throwable> insertObjectExceptions) {
@@ -827,37 +903,51 @@ public final class GoogleCloudStorageGrpcWriteChannelTest {
       // update the list).
       this.insertObjectExceptions = Lists.newArrayList(insertObjectExceptions);
     }
+  }
 
-    private static class InsertRequestObserver implements StreamObserver<WriteObjectRequest> {
+  private static class InsertRequestObserver implements StreamObserver<WriteObjectRequest> {
 
-      private StreamObserver<WriteObjectResponse> responseObserver;
-      private Object object = DEFAULT_OBJECT;
-      WriteObjectResponse writeObjectResponse =
-          WriteObjectResponse.newBuilder().setResource(object).build();
-      Throwable insertRequestException;
-      boolean resumeFromInsertException = false;
+    private StreamObserver<WriteObjectResponse> responseObserver;
+    private Object object = DEFAULT_OBJECT;
+    WriteObjectResponse writeObjectResponse =
+        WriteObjectResponse.newBuilder().setResource(object).build();
+    Throwable insertRequestException;
+    boolean resumeFromInsertException = false;
+    boolean errored = false;
 
-      @Override
-      public void onNext(WriteObjectRequest request) {
-        if (insertRequestException != null) {
-          responseObserver.onError(insertRequestException);
-          if (resumeFromInsertException) {
-            insertRequestException = null;
-          }
-        } else {
-          responseObserver.onNext(writeObjectResponse);
+    public InsertRequestObserver(WriteObjectResponse response, Throwable throwable) {
+      if (response != null) {
+        writeObjectResponse = response;
+      }
+      if (throwable != null) {
+        insertRequestException = throwable;
+      }
+    }
+
+    @Override
+    public void onNext(WriteObjectRequest request) {
+      if (insertRequestException != null) {
+        onError(insertRequestException);
+        if (resumeFromInsertException) {
+          insertRequestException = null;
         }
       }
+    }
 
-      @Override
-      public void onError(Throwable t) {
-        responseObserver.onError(t);
-      }
+    @Override
+    public void onError(Throwable t) {
+      errored = true;
+      responseObserver.onError(t);
+    }
 
-      @Override
-      public void onCompleted() {
-        responseObserver.onCompleted();
-      }
+    @Override
+    public void onCompleted() {
+      responseObserver.onNext(writeObjectResponse);
+      responseObserver.onCompleted();
+    }
+
+    public boolean isErrored() {
+      return errored;
     }
   }
 }