+++ gcs/CHANGES.md
@@ -2,6 +2,22 @@
 
 ## Next
 
+
+1.  Added dependency on new storage client library: [google-cloud-storage](https://github.com/googleapis/java-storage/tree/main/google-cloud-storage).
+
+1. Add support for OAuth2 based client authentication that retrieves
+a refresh token using the authorisation code grant flow configured via the
+following properties:
+
+    ```
+    fs.gs.auth.client.id
+    fs.gs.auth.client.secret
+    fs.gs.auth.refresh.token
+   ```
+## 2.2.10 - 2022-12-11
+
+1. Update dependencies to latest version
+
 ## 2.2.9 - 2022-11-09
 
 1.  The Google Cloud Storage Connector now can be used as a